split read_some_rows & finished testing
