--- include/boost/mysql/detail/network_algorithms/read_some_rows.hpp
@@ -1,75 +1,0 @@
-//
-// Copyright (c) 2019-2023 Ruben Perez Hidalgo (rubenperez038 at gmail dot com)
-//
-// Distributed under the Boost Software License, Version 1.0. (See accompanying
-// file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
-//
-
-#ifndef BOOST_MYSQL_DETAIL_NETWORK_ALGORITHMS_READ_SOME_ROWS_HPP
-#define BOOST_MYSQL_DETAIL_NETWORK_ALGORITHMS_READ_SOME_ROWS_HPP
-
-#include <boost/mysql/diagnostics.hpp>
-#include <boost/mysql/error_code.hpp>
-#include <boost/mysql/execution_state.hpp>
-#include <boost/mysql/rows.hpp>
-#include <boost/mysql/rows_view.hpp>
-
-#include <boost/mysql/detail/channel/channel.hpp>
-#include <boost/mysql/detail/execution_processor/execution_processor.hpp>
-#include <boost/mysql/detail/execution_processor/execution_state_impl.hpp>
-
-namespace boost {
-namespace mysql {
-namespace detail {
-
-// Note: it is OK depending on execution_state_impl, but
-// we want to avoid depending on the static one here, since it requires C++14 and Describe
-
-template <class Stream>
-rows_view read_some_rows(
-    channel<Stream>& channel,
-    execution_state_impl& st,
-    error_code& err,
-    diagnostics& diag
-);
-
-template <
-    class Stream,
-    BOOST_ASIO_COMPLETION_TOKEN_FOR(void(::boost::mysql::error_code, ::boost::mysql::rows_view))
-        CompletionToken>
-BOOST_ASIO_INITFN_AUTO_RESULT_TYPE(CompletionToken, void(error_code, rows_view))
-async_read_some_rows(
-    channel<Stream>& channel,
-    execution_state_impl& result,
-    diagnostics& diag,
-    CompletionToken&& token
-);
-
-template <class Stream>
-std::size_t read_some_rows(
-    channel<Stream>& channel,
-    execution_processor& st,
-    const output_ref& output,
-    error_code& err,
-    diagnostics& diag
-);
-
-template <
-    class Stream,
-    BOOST_ASIO_COMPLETION_TOKEN_FOR(void(::boost::mysql::error_code, std::size_t)) CompletionToken>
-BOOST_ASIO_INITFN_AUTO_RESULT_TYPE(CompletionToken, void(error_code, std::size_t))
-async_read_some_rows(
-    channel<Stream>& channel,
-    execution_processor& st,
-    const output_ref& output,
-    diagnostics& diag,
-    CompletionToken&& token
-);
-
-}  // namespace detail
-}  // namespace mysql
-}  // namespace boost
-
-#include <boost/mysql/detail/network_algorithms/impl/read_some_rows.hpp>
-
-#endif
