+++ test/unit/CMakeLists.txt
@@ -35,7 +35,8 @@ add_executable(
     detail/execution_processor/concepts.cpp
     detail/network_algorithms/read_resultset_head.cpp
     detail/network_algorithms/start_execution_impl.cpp
-    detail/network_algorithms/read_some_rows.cpp
+    detail/network_algorithms/read_some_rows_dynamic.cpp
+    detail/network_algorithms/read_some_rows_static.cpp
     detail/network_algorithms/execute_impl.cpp
     detail/network_algorithms/high_level_execution.cpp
     detail/network_algorithms/close_statement.cpp
